+++ src/env/types.ts
@@ -1,15 +1,10 @@
-import { GLProbe } from '../filters/GLProbes/GLProbe';
+import type { GLProbe } from '../filters/GLProbes/GLProbe';
 import type { DOMWindow } from 'jsdom';
 
-export type TCopyPasteData = {
-  copiedText?: string;
-  copiedStyle?: Record<string, string>;
-};
 export type TFabricEnv = {
   document: Document;
   window: Window | DOMWindow;
   isTouchSupported: boolean;
   WebGLProbe: GLProbe;
   dispose(element: Element): void;
-  copyPasteData: TCopyPasteData;
 };